add error handling

+++ src/controllers/products.controller.js
@@ -1,4 +1,5 @@
 const productsService = require('../services/products.service');
+const asyncController = require('../shared/asyncController');
 
 const findProductById = async (req, res) => {
   const { productId } = req.params;
@@ -9,13 +10,16 @@ const findProductById = async (req, res) => {
 };
 
 
-const findProducts = async (req, res) => {
-  const data = await productsService.findProducts();
-  return res.json(data);
+const findProducts = async (req, res, next) => {
+  try { 
+    const data = await productsService.findProducts();
+    return res.json(data);
+  } catch (e) {
+    next(e);
+  }
 };
 
-
-const createProduct = async (req, res) => {
+const createProduct = asyncController(async (req, res, ) => {
   const { title, amount, price, expirationDate } = req.body;
 
   const { userId } = req.user;
@@ -27,14 +31,15 @@ const createProduct = async (req, res) => {
   return res.json({
     message: result,
   });
-};
+});
 
 const updateProduct = async (req, res) => {
   const { id } = req.params;
   const { title, amount, price } = req.body;
+  const { userId } = req.user;
 
   const result = await productsService.updateProduct({
-    id, title, amount, price
+    id, title, amount, price, userId
   });
 
   return res.json({ message: result });
@@ -43,8 +48,9 @@ const updateProduct = async (req, res) => {
 
 const deleteProduct = async (req, res) => {
   const { productId } = req.params;
+  const { userId } = req.user;
 
-  const result = await productsService.deleteProduct(productId);
+  const result = await productsService.deleteProduct({productId, userId});
 
   return res.status(404).json({
     message: result,
